* Palettes/2Controls/GormFormAttributesInspector.m           fix some format string warnings
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@37926 72102866-910b-0410-8b05-ffd578937521

+++ Palettes/2Controls/GormFormAttributesInspector.m
@@ -181,7 +181,7 @@ NSUInteger numberStepperValue;
 		}
 	      else
 		{
-		  [object addEntry: [NSString stringWithFormat: @"Field (%i)", rows]];
+		  [object addEntry: [NSString stringWithFormat: @"Field (%ld)", (long)rows]];
 		}
 	    }
 	  cell.height = (rect.size.height + inter.height) / fields - inter.height;
@@ -199,7 +199,7 @@ NSUInteger numberStepperValue;
 
       while(delta > 0)
 	{
-	  [object addEntry: [NSString stringWithFormat: @"Field (%i)", rows]];
+	  [object addEntry: [NSString stringWithFormat: @"Field (%ld)", (long)rows]];
 	  delta--;
 	  rows++;
 	}
